Refactor appointment editing in client management
This commit refactors the appointment editing functionality in the client management module. The changes include displaying current appointment details for all clients with the same name and allowing the user to select which appointment to edit. Inputs for new appointment details such as date, time in, and duration are provided. The commit also calculates the new time out based on the new time in and duration. The code has been updated to use the selected row number instead of the cell row in the update section.